[WIP] Prepare pallet for extending aura
