+++ pallets/private-balances-aura-ext/Cargo.toml
@@ -1,17 +1,20 @@
 [package]
 edition = "2021"
 license = "GPLv3"
-name = "pallet-trusted-authorities"
+name = "pallet-private-balances-aura-ext"
 version = "0.1.0"
 
 [dependencies]
 # Substrate
 frame-support = { workspace = true }
 frame-system = { workspace = true }
+
 sp-core = { workspace = true }
 sp-io = { workspace = true }
 sp-runtime = { workspace = true }
 sp-std = { workspace = true }
+
+pallet-aura = { workspace = true }
 
 # Parity
 parity-scale-codec = { workspace = true }
