+++ runtime/common/construct_runtime.rs
@@ -27,7 +27,7 @@ macro_rules! construct_runtime {
 
 				Aura: pallet_aura = 25,
 				AuraExt: cumulus_pallet_aura_ext = 26,
-				TrustedAuthorities: pallet_trusted_authorities = 27,
+				PrivateBalancesAuraExt: pallet_private_balances_aura_ext = 27,
 
 				Balances: pallet_balances = 30,
